fix warning

+++ libraries/octree/src/OctreeSceneStats.cpp
@@ -773,7 +773,7 @@ void OctreeSceneStats::trackIncomingOctreePacket(ReceivedMessage& message, bool 
 
     // Guard against possible corrupted packets... with bad timestamps
     const qint64 MAX_RESONABLE_FLIGHT_TIME = 200 * USECS_PER_SECOND; // 200 seconds is more than enough time for a packet to arrive
-    const qint64 MIN_RESONABLE_FLIGHT_TIME = -1 * USECS_PER_SECOND; // more than 1 second of "reverse flight time" would be unreasonable
+    const qint64 MIN_RESONABLE_FLIGHT_TIME = -1 * (qint64)USECS_PER_SECOND; // more than 1 second of "reverse flight time" would be unreasonable
     if (flightTime > MAX_RESONABLE_FLIGHT_TIME || flightTime < MIN_RESONABLE_FLIGHT_TIME) {
         static QString repeatedMessage
             = LogHandler::getInstance().addRepeatedMessageRegex(
